Исправлено отображение сообщения по websocket в воркере нотификации
