+++ notification-worker/src/services/email.py
@@ -5,21 +5,11 @@ from email.mime.text import MIMEText
 from functools import lru_cache
 from typing import Annotated, Any
 
-import httpx
 from aiosmtplib import SMTP
-from core.config import settings
 from core.jinja2 import template_env
 from fastapi import Depends
 from integration.smtp import get_smtp_client
-from schemas.entity import UserInformation
-
-
-async def get_users_data(users_ids: list[uuid.UUID]) -> list[UserInformation]:
-    joined_users_ids = "&users_ids=".join([str(user_id) for user_id in users_ids])
-    url = f"{settings.auth_service_url}/?users_ids={joined_users_ids}"
-    async with httpx.AsyncClient() as client:
-        response = await client.get(url)
-        return [UserInformation(**data) for data in response.json()]
+from integration.http import get_users_data
 
 
 class BaseEmailSenderService:
